Working custom sort
