+++ src/app/settings/settings.html
@@ -66,8 +66,6 @@
       <div class="setting">
         <label for="itemSort" ng-i18next="Settings.SetSort"></label>
 
-        <sort-order-editor order="vm.itemSortOrder" on-sort-order-changed="vm.itemSortOrderChanged"></sort-order-editor>
-
         <div class="radioOptions">
           <label>
             <input type="radio" ng-model="vm.settings.itemSort" value="primaryStat">
@@ -97,7 +95,12 @@
             <input type="radio" ng-model="vm.settings.itemSort" value="typeThenName">
             <span ng-i18next="Settings.SortTypeThenName"></span>
           </label>
+          <label>
+            <input type="radio" ng-model="vm.settings.itemSort" value="custom">
+            <span ng-i18next="Settings.SortCustom"></span>
+          </label>
         </div>
+        <sort-order-editor ng-if="vm.settings.itemSort == 'custom'" order="vm.itemSortCustom" on-sort-order-changed="vm.itemSortOrderChanged"></sort-order-editor>
         <div class="fineprint" ng-i18next="Settings.DontForgetDupes"></div>
       </div>
     </section>
