+++ src/app/settings/settings.scss
@@ -70,6 +70,10 @@
     }
   }
 
+  .sort-order-editor {
+    margin-left: 2em;
+  }
+
   td:nth-child(2) {
     white-space: nowrap;
     min-width: 20em;
