+++ src/locale/dim.json
@@ -511,7 +511,7 @@
     "ColumnSizeAuto": "Auto",
     "DIMPopupsReset": "Reset previously hidden info tips",
     "Data": "Data",
-    "DontForgetDupes": "Don't forget you can search is:dupe or is:dupelower to quickly find duplicate items.",
+    "DontForgetDupes": "Don't forget you can search is:dupe or is:dupelower to quickly find duplicate items, and you can use the comparison tool to evaluate related items.",
     "EnableAdvancedStats": "Show stat quality rating on armor (D1)",
     "ExportSS": "Inventory spreadsheets",
     "ExportSSHelp": "Download a CSV list of your items that can be easily viewed in the spreadsheet app of your choice.",
@@ -535,11 +535,12 @@
     "ShowOverlay": "Show new items with a shiny overlay",
     "ShowReviews": "Show community ratings on weapons",
     "SizeItem": "Item size",
+    "SortCustom": "Custom Sort",
     "SortBasePower": "Base power (D2 only), then primary stat",
     "SortName": "Name",
     "SortPrimary": "Primary stat (attack or defense)",
     "SortRarity": "Rarity, then primary stat",
-    "SortRoll": "Quality rating (D1 armor only)",
+    "SortRoll": "Community rating or armor quality",
     "SortTypeThenName": "Item type then name",
     "SortTypeThenPrimary": "Item type then primary stat",
     "VaultColumns": "Vault maximum inventory width"
